NAV-41 Stubs for navigator API.
